--- a/audiofile/core/utils.py
+++ b/audiofile/core/utils.py
@@ -14,7 +14,7 @@ MAX_CHANNELS = {
 }
 r"""Maximum number of channels per format."""
 
-SNDFORMATS = ['wav', 'flac', 'ogg']
+SNDFORMATS = ['wav', 'flac', 'mp3', 'ogg']
 r"""File formats handled by soundfile"""