Closes #119

This adds reading and metadata support for MP3 files using ~~`soundfile >=0.11.0`~~ `soundfile>=0.12.1`.
I have not added writing support for MP3 files ~~as this depends on the system wide installation of `libsndfile` (compare https://github.com/bastibe/python-soundfile/issues/348), whereas for reading the `libsndfile` library bundled with `soundfile` is used (hence the dependency on ~~`soundfile>=0.11.0`~~ `soundfile>=0.12.1`).~~
If we think writing support for MP3 is needed as well, I would propose to add it in another pull request.

It also updates the benchmark page with the new results.
For the "Reading files" benchmark it presents now all results in a single figure. For the "Accessing metadata" benchmark,
results for MP4 are still shown in an extra figure as it takes much longer than for the other formats.

The good news is that the proposed changes make processing of MP3 faster (results as average per file):

| benchmark | main | this branch |
| --- | --- | --- |
| Reading files | 0.07 s | 0.03 s |
| Accessing metadata | 0.19 s | 0.0002 s |

**Benchmark page**

![image](https://github.com/audeering/audiofile/assets/173624/ad1f0291-3619-4f98-a8f5-ff7cbb9e2646)

/cc @frankenjoe