@hagenw 2024-01-22
> Hey, thanks for the update, unfortunately, I'm not quite sure what the update does. The main change is that we require a newer soundfiles version and that all tests and benchmarks are updated accordingly, right?

Yes, we require `soundfile>=0.12.1` as this comes with a pre-build `libsndfile` that does support reading of MP3 files (which was introduced in `libsndfile` in version `1.1.0`).
BTW, you can check for the `libsndfile` that is used by `soundfile` with:
```python
import soundfile

soundfile.__libsndfile_version__
```

`soundfile` 0.12.1 returns `'1.2.0'`.

In addition, it updates the tests as MP3 was used in some tests to represent an audio format not supported by `soundfile`, which is no longer true. In those cases I switched the test to MP4.

And the most files were changed as part of the updated benchmark pages, which now include the results for the changes of this pull request.