@schruefer 2024-01-22
Okay, great, makes a lot of sense to me!
I was just a little bit confused because I was looking for some bigger code changes to use `libsndfile`.
Sounds good to me!